[packages/template,core][s]: using nav from core
