+++ packages/core/src/ui/index.ts
@@ -1,2 +1,1 @@
 export { Nav } from "./Nav";
-export * from "./Icons";
